File handling rework
The application only moved over mod files of which the same name existed
in the CookedPC folder. I thought the game client wouldn't use/load
files that weren't already present since install. After doing a test
with another mod today I realized this was wrong and decided to rework
the way it does things.

Now the application moves over all files in the mods folder and keeps a
JSON list of what it moved so it can do a clean reversal of the
installation. And cleaned up the code a tiny bit.